+++ src/app/files/components/Toolbar.tsx
@@ -57,6 +57,7 @@ export function Toolbar({
             return (
               <ToolbarButton
                 key={action}
+                action={action}
                 {...buttonProps}
                 onClick={() => onAction(action)}
                 active={activeActions.has(action)}